fix!: Remove infinitely recursing `From` impl
BREAKING CHANGE: Remove the infinitely recursing implementation of
`From<&Box<[u8]>>` for `RedisValue`. In practice, this impl would
cause infinite recursion, so it was probably never used. Since this
is a change to a public API, it is technically a breaking change.

To convert a `&Box<[u8]>` to a `RedisValue`, users can clone the box.